Fix user settings page and remove billing settings
When selecting a tab in the list on the left, that option would disappear, and screw with the list. (Pressing "My account" would select it, but make the button vanish and replace it with the "Profiles" button). This fixes that.

It also removes all the billing info buttons. The rationale behind this is if you're using Discord adblock, chances are you don't want Discord handling your billing info in the first place. Should someone need to remove their billing info, they can temporarily disable their custom CSS.