Remove Super Reaction button next to existing reactions
When a post has reactions added to it already, there will be buttons to add another reaction, or a super reaction. The Super Reaction button is always the last child though, so using nth-last-child(1), this removes the super reaction button, and leaves the normal reaction button as well as existing reactions intact.